showing the connection list on setup page
