+++ pages/setup/[token].tsx
@@ -1,5 +1,6 @@
 import type { NextPage } from 'next';
 import AddEdit from '@components/connection/AddEdit';
+import ConnectionList from '@components/connection/ConnectionList';
 import useSWR from 'swr';
 import { fetcher } from '@lib/ui/utils';
 import { useRouter } from 'next/router';
@@ -7,16 +8,26 @@ import { useRouter } from 'next/router';
 const Setup: NextPage = () => {
   const router = useRouter();
   const { token } = router.query;
-  const { data: setup } = useSWR<any>([`/api/setup/${token}`], fetcher, { revalidateOnFocus: false });
+  const { data: setup } = useSWR<any>(token ? `/api/setup/${token}` : null, fetcher, { revalidateOnFocus: false });
   if (!token || !setup) {
     return null;
   } else {
     console.log(setup.data);
       switch(setup.data.path) {
         case '/admin/connection/new':
-          return <AddEdit setup={{...setup.data, token}} />;
+          return (
+            <>
+              <ConnectionList setupToken={token as string} />
+              <AddEdit setup={{ ...setup.data, token }}  />
+            </>
+          );
         default:
-          return <AddEdit setup={{...setup.data, token}} />;
+          return (
+            <>
+              <ConnectionList setupToken={token as string} />
+              <AddEdit setup={{ ...setup.data, token }} />
+            </>
+          );
       }
   }
 };
